Update proxy pkg
